organize and structure tests
